tests/wpt/tests/workers/constructors/Worker/use-base-url.html

<!doctype html>
<meta charset="utf-8">
<title>Use the document base url when resolving worker URLs</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<base href="/workers/constructors/Worker/sample_worker/">
<script>
    async_test(function(t) {
        var worker = new Worker('worker.js');
        var data = "foo";
        worker.postMessage(data);
        worker.onmessage = t.step_func_done(function(event) {
            assert_equals(event.data, data, "event.data does not match expected data");
        });
        worker.onerror = t.unreached_func("received error event");
    }, "Use the document base url when resolving worker URLs");
</script>
